% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantsBDM}
\alias{plantsBDM}
\title{Two surveys of Swiss plant communities from Swiss Biodiversity Monitoring.}
\format{A list with xx elements:
\describe{
  \item{y}{Array with presence/absence observed in 362 squares for 1733 species 
  during two surveys.}
  \item{elevation}{Vecotor with elevation (m) of each square.}
  
  
}}
\source{
\url{http://www.biodiversitymonitoring.ch/}
}
\usage{
plantsBDM
}
\description{
The Swiss biodiversity monitoring (BDM) has among others monitored the presence/absence 
of plant species at 1x1km squares. The data-set contains the data from 363 squares with exactly
two surveys conducted by professional botanists between 2010 and 2014.
}
\keyword{datasets}
