% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traitmat}
\alias{traitmat}
\title{Traits for the 1733 species observed during the BDM surveys.}
\format{A data.frame with the following collumns: \describe{ 
  \item{sla}{Specific leaf area.} \item{ch}{Canopy height.} \item{sm}{Seed
  mass.}
  
  
  }}
\source{
\url{http://www.uni-oldenburg.de/en/biology/landeco/research/projects/leda/}
}
\usage{
traitmat
}
\description{
The Swiss
}
\references{
Kleyer, M.; Bekker, R. M.; Knevel, I. C.; Bakker, J. P.;
  Thompson, K.; Sonnenschein, M., . . . Peco, B., 2008. The LEDA traitbase: a
  database of life-history traits of the Northwest European flora. - Journal
  of Ecology 96: 1266-1274.
}
\keyword{datasets}
